Move error files into confirmation folder
