+++ src/pages/Confirmation/__tests__/AccountAlreadyActive.test.jsx
