+++ src/pages/Confirmation/__tests__/FormConfirmationPage.test.jsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
-import FormConfirmationPage from './FormConfirmationPage';
+import FormConfirmationPage from '../FormConfirmationPage';
 
 const mockUseLocationState = { state: {} };
 const mockedUsedNavigate = jest.fn();
